chore(sweep): remove sweep.service v1 and promote v2 as the canonical implementation (#628)
